provide a method for plugins to expand all jinja2 templates
Provide a method that allows plugins, custom or not, to expand all
Jinja2 templates contained in variables by calling a method. The method
would accept dictionary as an argument which may or may not contain
Jinja2 templates as values in the dictionary and would return another
dictionary that containing same variables but with the Jinja2 templates
expanded.

Signed-off-by: Rishabh Dave <ridave@redhat.com>